Refactor IAM handling
